// devtools/projects/ng-devtools-backend/src/lib/component-tree/core-enums.d.ts
/**
 * @license
 * Copyright Google LLC All Rights Reserved.
 *
 * Use of this source code is governed by an MIT-style license that can be
 * found in the LICENSE file at https://angular.dev/license
 */
export declare enum ChangeDetectionStrategy {
    OnPush = 0,
    Default = 1
}
export declare enum AcxChangeDetectionStrategy {
    Default = 0,
    OnPush = 1
}
export declare enum Framework {
    Angular = "angular",
    ACX = "acx",
    Wiz = "wiz"
}